Reorganized the package system
