--- src/main/java/com/codecool/shop/dao/implementation/DB/ProductCategoryDaoDB.java
@@ -1,5 +1,0 @@
-package com.codecool.shop.dao.implementation.DB;
-
-
-public class ProductCategoryDaoDB {
-}
