+++ src/main/java/com/codecool/shop/dao/implementation/db/OrderDaoDB.java
@@ -1,0 +1,5 @@
+package com.codecool.shop.dao.implementation.db;
+
+
+public class OrderDaoDB {
+}
