+++ src/main/java/com/codecool/shop/dao/implementation/db/ProductDaoDB.java
@@ -1,0 +1,5 @@
+package com.codecool.shop.dao.implementation.db;
+
+
+public class ProductDaoDB {
+}
