+++ src/main/java/com/codecool/shop/dao/implementation/mem/OrderDaoMem.java
@@ -1,4 +1,4 @@
-package com.codecool.shop.dao.implementation.Mem;
+package com.codecool.shop.dao.implementation.mem;
 
 import com.codecool.shop.dao.OrderDao;
 import com.codecool.shop.model.Order;
